+++ src/main/java/com/example/earthtalk/domain/news/config/NewsConfig.java
@@ -1,0 +1,11 @@
+package com.example.earthtalk.domain.news.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableConfigurationProperties(CrawlConfig.class)
+public class NewsConfig {
+
+}
